+++ magik-checks/src/test/java/org/stevenlooman/sw/magik/checks/ExemplarSlotCountTest.java
@@ -33,4 +33,15 @@ public class ExemplarSlotCountTest extends MagikCheckTestBase {
     assertThat(issues).isEmpty();
   }
 
+  @Test
+  public void testVec() {
+    ExemplarSlotCountCheck check = new ExemplarSlotCountCheck();
+    check.maxSlotCount = 2;
+    String code =
+        "def_slotted_exemplar(:exemplar, vec(" +
+            "vec(:slot_1, _unset)))\n";
+    List<MagikIssue> issues = runCheck(code, check);
+    assertThat(issues).isEmpty();
+  }
+
 }