+++ magik-lint/src/main/java/org/stevenlooman/sw/magik/lint/MagikLint.java
@@ -24,6 +24,7 @@ import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.HashMap;
@@ -74,6 +75,12 @@ public class MagikLint {
     options.addOption(Option.builder()
         .longOpt("watch")
         .desc("Watch the given directory/file for changes")
+        .build());
+    options.addOption(Option.builder()
+        .longOpt("max-infractions")
+        .desc("Set max number of reporter infractions")
+        .hasArg()
+        .type(PatternOptionBuilder.NUMBER_VALUE)
         .build());
   }
 
@@ -185,6 +192,22 @@ public class MagikLint {
     }
   }
 
+  private boolean isMagikIssueDisabled(
+      MagikIssue magikIssue, InstructionsHandler instructionsHandler) {
+    int line = magikIssue.line();
+    int column = magikIssue.column();
+    String checkKey = magikIssue.check().getCheckKeyKebabCase();
+
+    Map<String, String> scopeInstructions =
+        instructionsHandler.getInstructionsInScope(line, column);
+    Map<String, String> lineInstructions =
+        instructionsHandler.getInstructionsAtLine(line);
+    String[] scopeDisableds = scopeInstructions.getOrDefault("disable", "").split(",");
+    String[] lineDisableds = lineInstructions.getOrDefault("disable", "").split(",");
+    return Arrays.asList(scopeDisableds).contains(checkKey)
+           || Arrays.asList(lineDisableds).contains(checkKey);
+  }
+
   /**
    * Check all files.
    *
@@ -197,14 +220,16 @@ public class MagikLint {
   private int checkFiles(Iterable<CheckInfo> checkInfos, Iterable<Path> paths)
       throws IOException, ParseException {
     int returnCode = 0;
+    int infractionCount = 0;
 
-    Comparator<CheckInfraction> byPath = Comparator.comparing(ci -> ci.getPath().toString());
     Comparator<CheckInfraction> byLine = Comparator.comparing(ci -> ci.getMagikIssue().line());
     Comparator<CheckInfraction> byColumn = Comparator.comparing(ci -> ci.getMagikIssue().column());
 
     Reporter reporter = getReporter();
     for (Path path : paths) {
+      System.out.println("Checking: " + path);
       MagikVisitorContext context = buildContext(path);
+      InstructionsHandler instructionsHandler = new InstructionsHandler(context);
       List<CheckInfraction> fileInfractions = new ArrayList<>();
 
       // run checks, report issues
@@ -215,6 +240,7 @@ public class MagikLint {
 
         List<MagikIssue> magikIssues = runCheck(context, checkInfo);
         List<CheckInfraction> checkInfractions = magikIssues.stream()
+            .filter(magikIssue -> !isMagikIssueDisabled(magikIssue, instructionsHandler))
             .map(magikIssue -> new CheckInfraction(path, checkInfo, magikIssue))
             .collect(Collectors.toList());
 
@@ -222,14 +248,19 @@ public class MagikLint {
       }
 
       fileInfractions.sort(
-          byPath
-          .thenComparing(byLine)
+          byLine
           .thenComparing(byColumn));
       for (CheckInfraction checkInfraction : fileInfractions) {
         reporter.reportIssue(checkInfraction);
 
         int checkReturnCode = getReturnCode(checkInfraction);
         returnCode = returnCode | checkReturnCode;
+
+        infractionCount++;
+        if (commandLine.hasOption("max-infractions")
+            && infractionCount == (Long)commandLine.getParsedOptionValue("max-infractions")) {
+          return returnCode;
+        }
       }
     }